search: move search to a custom hook
