+++ src/templates/home.js
@@ -1,9 +1,8 @@
-import React, { useState, useEffect, useMemo, useCallback } from 'react'
+import React, { useState, useMemo, useCallback } from 'react'
 import { useColorMode, Flex, Stack, IconButton, Button } from '@chakra-ui/core'
-import * as JSSearch from 'js-search'
-import SearchTermsSanitizer from '../utils/SearchTermsSanitizer'
 
 import { STATUS, useLoadMore } from '../hooks/useLoader'
+import { useSearch } from '../hooks/useSearch'
 
 import Layout from '../components/layout'
 import SEO from '../components/seo'
@@ -22,70 +21,31 @@ function sortByDate(order, list) {
   return list
 }
 
-function buildIndexes(data) {
-  const search = new JSSearch.Search('id')
-
-  search.indexStrategy = new JSSearch.PrefixIndexStrategy()
-  search.sanitizer = SearchTermsSanitizer
-  search.searchIndex = new JSSearch.TfIdfSearchIndex('id')
-
-  search.addIndex('locale')
-  search.addIndex('senority')
-  search.addIndex('stack')
-
-  search.addDocuments(data)
-
-  return search
-}
-
 const IndexPage = ({ pageContext }) => {
-  const { data, loadNextPage, pageInfo, status } = useLoadMore({
+  const { data, pageInfo, status, loadNextPage } = useLoadMore({
     initialData: pageContext.nodes,
     initialPageInfo: pageContext.pageInfo
   })
-
-  const [criteria, setCriteria] = useState('')
-  const [results, setResults] = useState([])
-  const [sort, setSort] = useState(() => 'desc')
-  const [engine, setEngine] = useState(null)
-
   const { colorMode, toggleColorMode } = useColorMode()
-
-  const backgroundButtonColor = { dark: 'blue.100', light: 'blue.200' }
-  const colorButton = { dark: 'black', light: 'white' }
-  const stateButton = { dark: 'blue.200', light: 'blue.100' }
-
-  useEffect(() => {
-    setEngine(() => buildIndexes(data))
-  }, [data])
-
-  useEffect(() => {
-    if (criteria.length > 2) {
-      const results = engine.search(criteria)
-      setCriteria(criteria)
-
-      setResults(() => results)
-    }
-
-    if (criteria.length <= 0) {
-      setResults(() => [])
-      setCriteria(() => '')
-    }
-  }, [criteria, engine])
-
-  const handleCriteriaChange = useCallback(
-    ({ target }) => setCriteria(target.value),
-    []
-  )
+  const { onSearch, results, term } = useSearch({ data })
+  const [sort, setSort] = useState(() => 'desc')
 
   const handleChangeSort = useCallback(
     ({ target }) => setSort(() => target.value),
     []
   )
 
-  const persons = useMemo(() => ((criteria.length > 2) ? results : data), [results, data, criteria])
+  const persons = useMemo(
+    () => ((term.length > 2) ? results : data),
+    [results, data, term]
+  )
 
   const isLoading = status === STATUS.loading
+  const shouldRenderLoadMoreButton = pageInfo.hasNextPage && term.length < 3
+
+  const backgroundButtonColor = { dark: 'blue.100', light: 'blue.200' }
+  const colorButton = { dark: 'black', light: 'white' }
+  const stateButton = { dark: 'blue.200', light: 'blue.100' }
 
   return (
     <Layout>
@@ -117,9 +77,9 @@ const IndexPage = ({ pageContext }) => {
 
           <Search
             flex={1}
-            onChange={handleCriteriaChange}
+            onChange={onSearch}
             onSortChange={handleChangeSort}
-            value={criteria}
+            value={term}
             sort={sort}
           />
 
@@ -130,7 +90,7 @@ const IndexPage = ({ pageContext }) => {
             ))}
           </List>
 
-          {(pageInfo.hasNextPage && criteria.length < 2) && (
+          {shouldRenderLoadMoreButton && (
             <Button
               borderRadius="lg"
               onClick={loadNextPage}
